backend/requirements.txt
Flask==2.0.3
Werkzeug==2.2.2
psycopg2-binary==2.9.9
